requirements.txt
python-dotenv
openai==1.34.0
twine
requests

# Langchain
langchain==0.2.5
langchain-core==0.2.7
langchain-openai==0.1.8
langchain-anthropic==0.1.15
langchain-community==0.2.5
langgraph==0.0.69
